DiscountView

+++ microkernel/plugins/discountplugin/src/main/java/br/edu/ifba/inf008/plugins/DiscountPolicy.java
@@ -1,7 +1,11 @@
 package br.edu.ifba.inf008.plugins;
 
+import br.edu.ifba.inf008.domain.Order;
+
+import java.math.BigDecimal;
+
 public interface DiscountPolicy {
     String getName();
 
-    Double calculateDiscount(Order order);
+    BigDecimal calculateDiscount(Order order);
 }
